Leveldb lib directory move
Moved lib under the scripts directory.
Substituted prints for logfunc.
