+++ .gitignore
@@ -6,3 +6,4 @@ scripts/.DS_Store
 scripts/.DS_Store
 .DS_Store
 .DS_Store
+.DS_Store
