+++ scripts/artifacts/LocalStorage.py
@@ -63,7 +63,7 @@ def parse_ls_ldb_record(record):
                 parsed['key'] = parsed['key'].lstrip(b'\x00').decode('utf-16')
 
         except Exception as e:
-            print ("Origin/key parsing error: {}".format(e))
+            logfunc(str("Origin/key parsing error: {}".format(e)))
             return
 
         try:
@@ -80,7 +80,7 @@ def parse_ls_ldb_record(record):
                 parsed['value'] = ''
 
         except Exception as e:
-            print (f'Value parsing error: {e}')
+            logfunc(str(f'Value parsing error: {e}'))
             return
 
     for item in parsed.values():
@@ -93,17 +93,17 @@ def get_local_storage(ls_path):
     ''' This code was taken from the file utils.py from Ryan Benson's Hindsight project '''
     results = []
 
-#    print ('Local Storage:')
-#    print (f' - Reading from {ls_path}')
+#    logfunc ('Local Storage:')
+#    logfunc (f' - Reading from {ls_path}')
 
     local_storage_listing = os.listdir(ls_path)
-#    print (f' - {len(local_storage_listing)} files in Local Storage directory')
+#    logfunc (f' - {len(local_storage_listing)} files in Local Storage directory')
     filtered_listing = []
 
     # Chrome v61+ used leveldb for LocalStorage, but kept old SQLite .localstorage files if upgraded.
     ls_ldb_path = ls_path
     ls_ldb_records = get_ldb_records(ls_ldb_path)
-#    print (f' - Reading {len(ls_ldb_records)} Local Storage raw LevelDB records; beginning parsing')
+#    logfunc (f' - Reading {len(ls_ldb_records)} Local Storage raw LevelDB records; beginning parsing')
     for record in ls_ldb_records:
         ls_item = parse_ls_ldb_record(record)
         if ls_item and ls_item.get('record_type') == 'entry':
@@ -112,7 +112,7 @@ def get_local_storage(ls_path):
                             ls_item['seq'], ls_item['state'], str(ls_item['origin_file'])))
 
 #    self.artifacts_counts['Local Storage'] = len(results)
-#    print (f' - Parsed {len(results)} items from {len(filtered_listing)} files')
+#    logfunc (f' - Parsed {len(results)} items from {len(filtered_listing)} files')
 #    self.parsed_storage.extend(results)
     return results
