+++ scripts/lib/ccl_chrome_indexeddb/.gitignore
