+++ scripts/lib/ccl_chrome_indexeddb/ccl_simplesnappy.py
