+++ scripts/lib/ccl_chrome_indexeddb/dump_indexeddb_details.py
